fix: 修复图片读取

+++ src/main/fs.js
@@ -4,11 +4,16 @@ export const checkAneMoveFile = (filepath) => {
     console.log('Check file...')
     if (fs.existsSync(filepath + 'data.db') == false) {
         console.log('Moving file...')
-        const files = ['data.db', 'config.json', 'icon.png']
+        const files = ['data.db', 'config.json', 'config.example.json', 'icon.png']
         for (const file of files) {
-            fs.renameSync(filepath + `app.asar.unpacked/resources/${file}`, filepath + `${file}`)
+            if (fs.existsSync(filepath + `app.asar.unpacked/resources/${file}`, filepath + `${file}`)) {
+                fs.renameSync(filepath + `app.asar.unpacked/resources/${file}`, filepath + `${file}`)
+            }
         }
+        console.log('The avatar path is: ' + filepath + `app.asar.unpacked/resources/image/default-avatar.jpg`)
+        fs.mkdirSync(filepath + `image`)
         fs.renameSync(filepath + `app.asar.unpacked/resources/image/default-avatar.jpg`, filepath + `image/default-avatar.jpg`)
+        fs.rmdirSync(filepath + 'app.asar.unpacked/resources/image')
         fs.rmdirSync(filepath + 'app.asar.unpacked/resources')
         fs.rmdirSync(filepath + 'app.asar.unpacked')
         console.log('Move file success!')
@@ -78,8 +83,12 @@ export const setLaunchAtLogin = (path, enable) => {
 }
 
 export const copyAvatarToResources = (path, old) => {
-    if (fs.existsSync(resources_path + '/image/' + old)) {
+    if (old !== 'default-avatar.jpg' && fs.existsSync(resources_path + '/image/' + old)) {
         fs.rmSync(resources_path + '/image/' + old)
     }
     fs.copyFileSync(path, resources_path + '/image/' + path.split('\\')[path.split('\\').length - 1])
+}
+
+export const readImageByBase64 = (path) => {
+    return 'data:image/png;base64,' +  fs.readFileSync(path).toString('base64')
 }
